Fixing more size_t -> int16_t (and similar) warnings

+++ hpx/lcos/local/futures_factory.hpp
@@ -201,10 +201,12 @@ namespace hpx { namespace lcos { namespace local
                     return threads::register_thread_nullary(
                         util::deferred_call(
                             &base_type::run_impl, std::move(this_)),
-                        util::thread_description(this->f_, "task_object::apply"),
+                        util::thread_description(
+                            this->f_, "task_object::apply"),
                         threads::pending_do_not_schedule, true,
                         threads::thread_priority_boost,
-                        threads::thread_schedule_hint(get_worker_thread_num()),
+                        threads::thread_schedule_hint(
+                            static_cast<std::int16_t>(get_worker_thread_num())),
                         stacksize, ec);
                 }
                 else {
